Move to a dedicated renderer struct to keep the API clean
