+++ pkg/changelog/builder.go
@@ -3,14 +3,10 @@ package changelog
 import (
 	"context"
 	"fmt"
-	"net/http"
-	"regexp"
-	"strconv"
 	"strings"
 
 	"github.com/grafana/grafana-github-actions-go/pkg/ghgql"
 	"github.com/grafana/grafana-github-actions-go/pkg/toolkit"
-	"github.com/rs/zerolog"
 
 	"github.com/google/go-github/v50/github"
 )
@@ -76,211 +72,6 @@ type ChangelogBody struct {
 	Features           []ghgql.PullRequest
 }
 
-func (body *ChangelogBody) ToMarkdown(tk *toolkit.Toolkit) string {
-	out := strings.Builder{}
-	out.WriteString("# ")
-	out.WriteString(body.Version)
-	if body.ReleaseDate != "" {
-		out.WriteString(" (")
-		out.WriteString(body.ReleaseDate)
-		out.WriteString(")")
-	}
-	out.WriteString("\n\n")
-	if len(body.Features) > 0 {
-		out.WriteString("### Features and enhancements\n\n")
-		writeIssueLines(&out, tk, body.Features)
-		out.WriteString("\n")
-	}
-	if len(body.Bugfixes) > 0 {
-		out.WriteString("### Bug fixes\n\n")
-		writeIssueLines(&out, tk, body.Bugfixes)
-		out.WriteString("\n")
-	}
-	if len(body.BreakingChanges) > 0 {
-		out.WriteString("### Breaking changes\n\n")
-		for _, notice := range body.BreakingChanges {
-			out.WriteString(notice)
-			out.WriteString("\n\n")
-		}
-	}
-	if len(body.DeprecationChanges) > 0 {
-		out.WriteString("### Deprecations\n\n")
-		for _, notice := range body.DeprecationChanges {
-			out.WriteString(notice)
-			out.WriteString("\n\n")
-		}
-	}
-	if len(body.PluginDevChanges) > 0 {
-		out.WriteString("### Plugin development fixes & changes\n\n")
-		writeIssueLines(&out, tk, body.PluginDevChanges)
-		out.WriteString("\n")
-	}
-	return out.String()
-}
-
-func writeIssueLines(out *strings.Builder, tk *toolkit.Toolkit, issues []ghgql.PullRequest) {
-	for _, issue := range issues {
-		out.WriteString(issueAsMarkdown(issue, tk))
-	}
-}
-
-var titleHeadlinePattern = regexp.MustCompile(`^([^:]*:)`)
-
-func issueAsMarkdown(issue ghgql.PullRequest, tk *toolkit.Toolkit) string {
-	ctx := context.Background()
-	out := strings.Builder{}
-
-	title := issue.GetTitle()
-	title = stripReleaseStreamPrefix(title)
-	title = titleHeadlinePattern.ReplaceAllString(title, "**$1**")
-	title = strings.TrimSuffix(title, ".")
-
-	out.WriteString("- ")
-	out.WriteString(title)
-	if issueHasLabel(issue, LabelEnterprise) || issue.GetRepoName() == "grafana-enterprise" {
-		out.WriteString(". (Enterprise)")
-	} else {
-		out.WriteString(". ")
-		out.WriteString(getIssueLink(issue))
-		if issue.GetAuthorLogin() != "" {
-			userLink, err := getUserLink(ctx, issue, tk)
-			if err != nil {
-			} else {
-				out.WriteString(", ")
-				out.WriteString(userLink)
-			}
-		}
-	}
-	out.WriteString("\n")
-	return out.String()
-}
-
-var releaseStreamPrefixPattern = regexp.MustCompile(`^(\[[^]]+\]:?) (.*)$`)
-
-func stripReleaseStreamPrefix(input string) string {
-	if releaseStreamPrefixPattern.MatchString(input) {
-		result := releaseStreamPrefixPattern.FindStringSubmatch(input)
-		return result[2]
-	}
-	return input
-}
-
-func getIssueLink(issue ghgql.PullRequest) string {
-	num := strconv.Itoa(issue.GetNumber())
-	out := strings.Builder{}
-	out.WriteString("[#")
-	out.WriteString(num)
-	out.WriteString("]")
-	out.WriteString("(https://github.com/grafana/grafana/issues/")
-	out.WriteString(num)
-	out.WriteString(")")
-	return out.String()
-}
-
-func isBotUser(login string) bool {
-	switch login {
-	case "grafanabot":
-		return true
-	default:
-		return false
-	}
-}
-
-func getPRNumberFromBackportBranch(ref string) (int, error) {
-	pat := regexp.MustCompile("^backport-(\\d+)-to-v\\d+\\.\\d+\\.x$")
-	match := pat.FindStringSubmatch(ref)
-	if len(match) < 1 {
-		return -1, fmt.Errorf("no number found in ref")
-	}
-	result, err := strconv.ParseInt(match[1], 10, 64)
-	return int(result), err
-
-}
-
-func getUserLink(ctx context.Context, issue ghgql.PullRequest, tk *toolkit.Toolkit) (string, error) {
-	logger := zerolog.Ctx(ctx)
-	user := issue.GetAuthorLogin()
-	if isBotUser(user) {
-		logger.Info().Msgf("PR#%d created by bot. Fetching original author from %s", issue.GetNumber(), issue.GetHeadRefName())
-		// If this looks like a bot user, take the author of the original PR if
-		// available:
-		origPrNumber, err := getPRNumberFromBackportBranch(issue.GetHeadRefName())
-		if err != nil {
-			return "", err
-		}
-		origPR, _, err := tk.GitHubClient().PullRequests.Get(context.Background(), issue.GetRepoOwner(), issue.GetRepoName(), origPrNumber)
-		if err != nil {
-			return "", err
-		}
-		user = origPR.User.GetLogin()
-	}
-	out := strings.Builder{}
-	out.WriteString("[@")
-	out.WriteString(user)
-	out.WriteString("]")
-	out.WriteString("(https://github.com/")
-	out.WriteString(user)
-	out.WriteString(")")
-	return out.String(), nil
-}
-
-func issueHasLabel(issue ghgql.PullRequest, label string) bool {
-	for _, l := range issue.Labels {
-		if l == label {
-			return true
-		}
-	}
-	return false
-}
-
-func getBreakingChangeNotice(issue ghgql.PullRequest) string {
-	return getNotice(issue, "Release notice breaking change")
-}
-
-func getDeprecationNotice(issue ghgql.PullRequest) string {
-	return getNotice(issue, "Deprecation notice")
-}
-
-func getNotice(issue ghgql.PullRequest, sectionStart string) string {
-	lines := strings.Split(issue.GetBody(), "\n")
-	startFound := false
-	result := strings.Builder{}
-	lastLine := len(lines) - 1
-	for idx, line := range lines {
-		if startFound {
-			l := strings.TrimSpace(line)
-			if result.Len() > 0 {
-				result.WriteString("\n")
-			} else {
-				// If there is a blank line right after the start, let's skip it:
-				if l == "" {
-					continue
-				}
-			}
-			result.WriteString(l)
-			if idx == lastLine {
-				// Trim tailing whitespaces before finalizing the output:
-				output := strings.Builder{}
-				output.WriteString(strings.TrimSpace(result.String()))
-				result = strings.Builder{}
-				result.WriteString(output.String())
-
-				if strings.HasSuffix(output.String(), "```") {
-					result.WriteString("\n")
-				} else {
-					result.WriteString(" ")
-				}
-				result.WriteString("Issue ")
-				result.WriteString(getIssueLink(issue))
-			}
-		}
-		if strings.Contains(line, sectionStart) {
-			startFound = true
-		}
-	}
-	return result.String()
-}
-
 func newChangelogBody() *ChangelogBody {
 	return &ChangelogBody{
 		DeprecationChanges: make([]string, 0, 10),
@@ -320,26 +111,6 @@ func isBug(issue ghgql.PullRequest) bool {
 		return true
 	}
 	return false
-}
-
-func getIssues(ctx context.Context, tk *toolkit.Toolkit, repo string, version string) ([]*github.Issue, error) {
-	result := make([]*github.Issue, 0, 10)
-	nextPage := 1
-	for nextPage > 0 {
-		opts := &github.SearchOptions{}
-		opts.Page = nextPage
-		tk.IncrRequestCount()
-		issues, resp, err := tk.GitHubClient().Search.Issues(ctx, fmt.Sprintf(`repo:%s label:"add to changelog" is:pr is:closed milestone:%s`, repo, version), opts)
-		if err != nil {
-			return nil, err
-		}
-		if resp.StatusCode != http.StatusOK {
-			return nil, fmt.Errorf("unexpected status code from GitHub API: %s", resp.Status)
-		}
-		result = append(result, issues.Issues...)
-		nextPage = resp.NextPage
-	}
-	return result, nil
 }
 
 func getMilestone(ctx context.Context, tk *toolkit.Toolkit, repo string, version string) (*github.Milestone, error) {
