+++ update-changelog/main.go
@@ -88,6 +88,12 @@ func main() {
 		logger.Fatal().Err(err).Msg("Failed to build changelog")
 	}
 
+	renderer := changelog.NewRenderer(tk)
+	renderedMarkdown, err := renderer.Render(ctx, body)
+	if err != nil {
+		logger.Fatal().Err(err).Msg("Failed to render changelog to markdown")
+	}
+
 	if preview {
 		if changelogFile != "" {
 			input, err := os.Open(changelogFile)
@@ -95,18 +101,18 @@ func main() {
 				logger.Fatal().Err(err).Msg("Failed to open changelog file")
 			}
 			defer input.Close()
-			if err := changelog.UpdateFile(ctx, os.Stdout, input, body, tk); err != nil {
+			if err := changelog.UpdateFile(ctx, os.Stdout, input, renderedMarkdown, body); err != nil {
 				logger.Fatal().Err(err).Msg("Failed to update changelog file")
 			}
 		} else {
-			fmt.Println(body.ToMarkdown(tk))
+			fmt.Println(renderedMarkdown)
 		}
 		return
 	}
 	if changelogFile != "" {
 		logger.Info().Msgf("Updating %s", changelogFile)
 
-		if err := changelog.UpdateFileAtPath(ctx, changelogFile, body, tk); err != nil {
+		if err := changelog.UpdateFileAtPath(ctx, changelogFile, renderedMarkdown, body); err != nil {
 			logger.Fatal().Err(err).Msg("Failed to update changelog file")
 		}
 	}
@@ -151,7 +157,7 @@ func main() {
 				logger.Fatal().Err(err).Msg("Failed to switch to target branch")
 			}
 
-			if err := changelog.UpdateFileAtPath(ctx, filepath.Join(repositoryPath, "CHANGELOG.md"), body, tk); err != nil {
+			if err := changelog.UpdateFileAtPath(ctx, filepath.Join(repositoryPath, "CHANGELOG.md"), renderedMarkdown, body); err != nil {
 				logger.Fatal().Err(err).Msg("Failed to update changelog")
 			}
 
@@ -260,7 +266,7 @@ func main() {
 		)
 		if _, err := comm.CreateOrUpdatePost(ctx, community.PostInput{
 			Title:    fmt.Sprintf("Changelog: Updates in Grafana %s", body.Version),
-			Body:     body.ToMarkdown(tk),
+			Body:     renderedMarkdown,
 			Category: communityCategoryID,
 		}); err != nil {
 			logger.Fatal().Err(err).Msg("Failed to post to the forums")
